move avcC/a1op/a1lx box handling into AVIF specific file
