+++ libheif/file.h
@@ -21,6 +21,7 @@
 #ifndef LIBHEIF_FILE_H
 #define LIBHEIF_FILE_H
 
+#include "avif.h"
 #include "box.h"
 #include "hevc.h"
